Add www-root file serving support for the sake of WebGL.
This adds the backwards compatible API for serving files with the
web_video_server, similar to the ros_web_video server.

Tested with the depthcloud js using the vp8 codec.

Note: there is a huge string copy in there that I can't figure out
how to do it differently. Please improve :-)

Signed-off-by: Peter Soetens <peter@thesourceworks.com>